Created a new file to set the View of LibraryList
Set the back button into white color.
