+++ Studioframe/ContentView.swift
@@ -39,7 +39,6 @@ struct ContentView: View {
                         .rotationEffect(isItemsMenuOpen ? .degrees(90) : .degrees(0))
                         .animation(.easeIn(duration: 0.5), value: isItemsMenuOpen)
                         
-                        
                         Spacer()
                     }
                 }
@@ -47,91 +46,6 @@ struct ContentView: View {
             
         }
         
-        
-    }
-}
-
-
-
-struct LocalLibraryListView: View {
-    
-    let items: [String] = [
-        "Black wooden Chair",
-        "Stool",
-        "Slate",
-        "Black wooden Chaird",
-        "Stoolf",
-        "Slate",
-        "Black wooden Chairg",
-        "Stoolh",
-        "Slatea"
-    ]
-    
-    init() {
-        UITableView.appearance().backgroundColor = .black
-      
-        UITableView.appearance().contentInset.top = -45
-    }
-    
-    
-    
-    
-    var body: some View {
-        
-        List {
-//            Section {
-                
-                ForEach(items, id: \.self) { item in
-                    HStack{
-                        Image("Image_Not_Available")
-                            .resizable()
-                            .scaledToFit()
-                        Spacer()
-                        VStack {
-                            Text(item)
-                                .foregroundColor(.white)
-                                .frame(width: 175, alignment: .topLeading)
-                                .font(.body.bold())
-
-                            Spacer()
-                        }
-                    }
-                    .padding(.top, 10)
-                }
-
-                
-                
-//            } header: {
-//               Text("Local Objects")
-//            }
-            
-//            Section {
-//
-//            } header: {
-//                Text("Online objects")
-//            }
-
-            
-            .listRowSeparator(.visible)
-            .listRowSeparatorTint(.white)
-            .listRowBackground(Color.clear)
-            .listRowInsets(EdgeInsets(top: 0, leading: 10, bottom: 15, trailing: 10))
-            
-        }
-        .padding(.leading, -50)
-        .navigationBarTitleDisplayMode(.inline)
-        .navigationTitle("Local Library")
-//        .toolbar {
-//            Button {
-//
-//            } label: {
-//                Text("+")
-//                    .foregroundColor(.white)
-//            }
-//
-//        }
-        
-  
         
     }
 }
@@ -177,7 +91,7 @@ struct ContentView_Previews : PreviewProvider {
     static var previews: some View {
         Group {
             NavigationView {
-            LocalLibraryListView()
+            ContentView()
                 .previewInterfaceOrientation(.portrait)
             }
         }
@@ -186,18 +100,3 @@ struct ContentView_Previews : PreviewProvider {
 #endif
 
 
-extension UINavigationController {
-    open override func viewDidLoad() {
-        let standardAppearance = UINavigationBarAppearance()
-        standardAppearance.configureWithTransparentBackground()
-        standardAppearance.backgroundColor = .black
-        standardAppearance.titleTextAttributes = [
-            .foregroundColor : UIColor.white
-        ]
-        
-        navigationBar.standardAppearance = standardAppearance
-        
-        
-        
-    }
-}
